requirements.txt
nextcord>=2.4.0
aiohttp>=3.8.0
python-dotenv>=0.20.0
Pillow>=9.0.0
numpy>=1.24.0
ffmpeg-python>=0.2.0